[DI] Decoupled DI container from flask app
- Flask app can now be created independently of DI container
- DI container has been broken up into more logical modules
